Double score... check point difficulty value methods

+++ Watch those Lines/DataManager.swift
@@ -83,7 +83,7 @@ class DataManager {
     
     // a simple function to add a new high score, to be called from your game logic
     // note that this doesn't sort or filter the scores in any way
-    func addNewScoreAndSave(newScore:Int, checkPoint: Int) {
+    func addNewScoreAndSave(newScore:Double, checkPoint: Int) {
         let newHighScore = BestScore(score: newScore, checkPoint: checkPoint);
         self.scores.append(newHighScore);
         self.saveScore();
